Template code nuked and fresh start-over
