--- config/index.ts
@@ -1,10 +1,0 @@
-import { Client } from "appwrite";
-import Constants from 'expo-constants';
-
-const client = new Client();
-
-client
-  .setEndpoint('https://cloud.appwrite.io/v1')
-  .setProject(Constants.expoConfig?.extra?.appwriteAppId as string ?? 'SAMPLE_ID');
-
-export default client
