+++ package.json
@@ -1,52 +1,23 @@
 {
   "name": "findmyfamily",
-  "main": "expo-router/entry",
-  "version": "0.0.1",
+  "version": "1.0.0",
+  "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "web": "expo start --web",
-    "test": "jest --watchAll"
-  },
-  "jest": {
-    "preset": "jest-expo"
+    "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/ngrok": "^4.1.0",
-    "@expo/vector-icons": "^13.0.0",
-    "@react-navigation/native": "^6.0.2",
-    "appwrite": "^13.0.0",
     "expo": "~49.0.15",
-    "expo-font": "~11.4.0",
-    "expo-linking": "~5.0.2",
-    "expo-router": "^2.0.0",
-    "expo-splash-screen": "~0.20.5",
     "expo-status-bar": "~1.6.0",
-    "expo-system-ui": "~2.4.0",
-    "expo-web-browser": "~12.3.2",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-native": "0.72.6",
-    "react-native-gesture-handler": "~2.12.0",
-    "react-native-safe-area-context": "4.6.3",
-    "react-native-screens": "~3.22.0",
-    "react-native-web": "~0.19.6"
+    "react-native": "0.72.6"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@types/react": "~18.2.14",
-    "@types/react-native": "^0.72.7",
-    "jest": "^29.2.1",
-    "jest-expo": "~49.0.0",
-    "react-test-renderer": "18.2.0",
     "typescript": "^5.1.3"
-  },
-  "overrides": {
-    "react-refresh": "~0.14.0"
-  },
-  "resolutions": {
-    "react-refresh": "~0.14.0"
   },
   "private": true
 }
